feat: migrate FormattedNumber to ts
